enhancement: split test file
